% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzlifetables.R
\docType{data}
\name{nzlifetables}
\alias{nzlifetables}
\title{New Zealand Cohort Life Tables}
\format{A data frame with 13,464 rows and 12 variables:
\itemize{
\item \code{sex}
\item \code{cohort}
\item \code{age} Exact age (years)
\item \code{lx} Number alive at exact age, out of 100,000 born
\item \code{Lx} Average number alive in the age interval, out of 100,000 born
\item \code{dx} Number dying in the age interval, out of 100,000 born
\item \code{px} Probability that a person who reaches this age lives another year
\item \code{qx} Probability that a person who reaches this age dies within a year
\item \code{mx} Central death rate for the age interval
\item \code{sx} Proportion of age group surviving another year
\item \code{ex} Expected number of years of life remaining at age x
\item \code{ex} Expected number of years of life remaining at age x
\item \code{projected} Whether or not the statistic is a projection
}}
\source{
http://www.stats.govt.nz/Census/2013-census/data-tables/electorate-tables.aspx
released under the CC BY 4.0 licence
https://creativecommons.org/licenses/by/4.0/
}
\usage{
nzlifetables
}
\description{
A dataset containing New Zealand Cohort Life Tables from Statistics New
Zealand
}
\keyword{datasets}
